scRUVIII

perform scRUVIII

scRUVIII(Y = Y, M = M, ctl = ctl, fullalpha = NULL, k = k,
  return.info = TRUE, cell_type = NULL, batch = NULL,
  return_all_RUV = TRUE, fast_svd = FALSE, rsvd_prop = 0.1)

Arguments

Y
The unnormalised SC data. A m by n matrix, where m is the number of observations and n is the number of features.

M
The replicate mapping matrix. The mapping matrix has m rows (one for each observation), and each column represents a set of replicates. The (i, j)-th entry of the mapping matrix is 1 if the i-th observation is in replicate set j, and 0 otherwise. See ruv::RUVIII for more details.

ctl
An index vector to specify the negative controls. Either a logical vector of length n or a vector of integers.

fullalpha
Not used. Please ignore.

k
The number of unwanted factors to remove. This is inherited from the ruvK argument from the scMerge::scMerge function.

return.info
Additional information relating to the computation of normalised matrix. We recommend setting this to true.

cell_type
An optional vector indicating the cell type information for each cell in the batch-combined matrix. If it is NULL, pseudo-replicate procedure will be run to identify cell type.

batch
Batch information inherited from the scMerge::scMerge function.

return_all_RUV
Whether to return extra information on the RUV function, inherited from the scMerge::scMerge function

fast_svd
If TRUE we will use the randomized SVD algorithm to speed up computation. Otherwise, we will use the standard SVD built into R.

rsvd_prop
If fast_svd = TRUE, then rsvd_prop is used to control for the accuracy of the randomized SVD computation. If a lower value is used on a lower dimensional data (say < 1000 cell) will potentially yield a less accurate computed result but with a gain in speed. The default of 0.1 tends to achieve a balance between speed and accuracy.