Create pseudo-replicate for scMerge

Create pseudo-replicate from single-cell RNA-seq data from different batches, experiments, and protocols.

scReplicate(sce, batch = NULL, kmeansK = NULL, exprs = "logcounts",
  hvg_exprs = "counts", marker = NULL, marker_list = NULL,
  replicate_prop = 1, cell_type = NULL, cell_type_match = FALSE,
  cell_type_inc = NULL, dist = "cor", WV = NULL, WV_marker = NULL,
  parallel = FALSE, parallelParam = bpparam(), return_all = FALSE)

Arguments

sce
A SingleCellExperiment object contains the batch-combined matrix with batch info in colData

batch
A vector indicates the batch information for each cell in the batch-combined matrix.

kmeansK
A vector indicates the kmeans's K for each batch, length of KmeansK needs to be the same as the number of batch.

exprs
A string inciates the assay that are used for batch correction, default is logcounts

hvg_exprs
A string inciates the assay that are used for highly variable genes identification, default is counts

marker
A vector of markers, which will be used in calcualtion of mutual nearest cluster. If no markers input, highly variable genes will be used instead

marker_list
A list of markers for each batch, which will be used in calcualtion of mutual nearest cluster.

replicate_prop
A number indicates the ratio of cells that are included in pseudo-replicates, ranges from 0 to 1

cell_type
A vector indicates the cell type information for each cell in the batch-combined matrix. If it is NULL, pseudo-replicate procedure will be run to identify cell type.

cell_type_match
Whether find mutual nearest cluster using cell type information

cell_type_inc
A vector indicates the indices of the cells that will be used to supervise the pseudo-replicate procedure

dist
The distance metrics that are used in the calcualtion of the mutual nearest cluster, default is Pearson correlation.

WV
A vector indicates the wanted variation factor other than cell type info, such as cell stages.

WV_marker
A vector indicates the markers of the wanted variation.

parallel
If TRUE, the algorithm will run in parallel.

parallelParam
The BiocParallelParam class is used.

return_all
If FALSE, only return the replicate matrix.

Value

If return_all is FALSE, return a replicate matrix.

If return_sce is TRUE, return the followings

repMat

replicate matrix

mnc

mutual nearest cluster

replicate vector

replicate vector

HVG

highly variable genes used in scReplicate