Combind several SingleCellExperiment objects from different batches/experiments

Combind several SingleCellExperiment objects from different batches/experiments.

sce_cbind(
  sce_list,
  method = "intersect",
  cut_off_batch = 0.01,
  cut_off_overall = 0.01,
  exprs = c("counts", "logcounts"),
  colData_names = NULL,
  batch_names = NULL
)

Arguments

sce_list
A list contains the SingleCellExperiment Object from each batch

method
A string indicates the method of combining the gene expression matrix, either union or intersect. Default to intersect. union only supports matrix class.

cut_off_batch
A numeric vector indicating the cut-off for the proportion of a gene is expressed within each batch

cut_off_overall
A numeric vector indicating the cut-off for the proportion of a gene is expressed overall data

exprs
A string vector indicating the expression matrices to be combined. The first assay named will be used to determine the proportion of zeores.

colData_names
A string vector indicating the colData that are combined

batch_names
A string vector indicating the batch names for the output sce object

Value

A SingleCellExperiment object with the list of SCE objects combined.

Author

Yingxin Lin

Examples

data('example_sce', package = 'scMerge')
batch_names<-unique(example_sce$batch)
sce_list<-list(example_sce[,example_sce$batch=='batch2'],
               example_sce[,example_sce$batch=='batch3'])
sce_combine<-sce_cbind(sce_list,batch_names=batch_names)

#> The assay named 'counts' will be used to determine the proportion of zeroes for each batch