Combind several SingleCellExperiment objects from different batches/experiments

Combind several SingleCellExperiment objects from different batches/experiments.

sce_cbind(sce_list, method = NULL, cut_off_batch = 0.01,
  cut_off_overall = 0.01, exprs = c("counts", "logcounts"),
  colData_names = NULL, batch_names = NULL)

Arguments

sce_list
A list contains the SingleCellExperiment Object from each batch

method
A string indicates the method of combining the gene expression matrix, either union or intersect

cut_off_batch
A numeric vector indicating the cut-off for the proportion of a gene is expressed within each batch

cut_off_overall
A numeric vector indicating the cut-off for the proportion of a gene is expressed overall data

exprs
A string vector indicating the expression matrices to be combined. The first assay named will be used to determine the proportion of zeores.

colData_names
A string vector indicating the colData that are combined

batch_names
A string vector indicating the batch names for the output sce object

Value

A SingleCellExperiment object with the list of SCE objects combined.

Examples

library(SingleCellExperiment)

#> Loading required package: SummarizedExperiment

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> Loading required package: BiocParallel

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

data('example_sce', package = 'scMerge')
batch_names<-unique(example_sce$batch)
sce_list<-list(example_sce[,example_sce$batch=='batch2'],
               example_sce[,example_sce$batch=='batch3'])
sce_combine<-sce_cbind(sce_list,batch_names=batch_names)

#> The assay named 'counts' will be used to determine the proportion of zeroes for each batch

#> As method is NULL, we will use the union method.